update_engine: new error message for update to rollback version
When an update is skipped because it corresponds to the rollback
version, an error message appears saying that the updates are disabled
by the administrator. This error message appears for both managed and
consumer users.

The expected behavior is to silently skip the update.

Currently this behavior is assigned to
ErrorCode::kOmahaUpdateIgnoredPerPolicy, which causes to display the
error message in the UI. As this error is also assigned to cases where
updates are actually blocked by admin policy, we need a new error
message specific for ignoring updates to rollback version.

Create kUpdateIgnoredRollbackVersion error code. When this error is
encountered, users should see the "No updates" message, so no further
error handling is required in Chromium.

BUG=b:246691369
TEST=1) FEATURES=test emerge-amd64-generic update_engine
     2a) Fake a preceding legacy rollback update.
     2b) Ensure "Your Chromebook is up to date" message appears for
     both managed and consumer accounts.

Change-Id: I3626298b34af12f39f57ad91c45169c34a36a073
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/4100471
Tested-by: Cristina Guerrero <crisguerrero@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Cristina Guerrero <crisguerrero@chromium.org>