update_engine: Support recovery key version
Parse the recovery key version and send within the request to Omaha
server. This is to distinguish between various recovery keys a device
might be using.

When a recovery key can't be read or is unparsable, the server will be
receiving an empty attribute like: recoverykeyversion=""
Otherwise, a positive integer should be parsed and set within the
`recoverykeyversion` attribute.

It is up to the server to decide what to do when the
`recoverykeyversion` field is missing and make a decision based off
additional fields in the request or simply not serve an update for the
apps that utilize this attribute.

```
INFO update_engine: [omaha_request_action.cc(273)] Request:
<?xml version="1.0" encoding="UTF-8"?>
<request requestid="6c5e5d4b-d3ca-44bc-8f33-ef1d7ed66879"
  sessionid="35891a80-d3c3-46c8-9721-c9a9f525360b" protocol="3.0"
  updater="ChromeOSUpdateEngine" updaterversion="0.1.0.0"
  installsource="ondemandupdate" ismachine="1" recoverykeyversion="1" >
...
</request>
```

BUG=b:267545513
TEST=emerge
TEST=deploy

Change-Id: I9344626c4e505af070a5e5b896652c29e5c0dd7f
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/4222216
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-by: Yuanpeng Ni‎ <yuanpengni@chromium.org>