update_engine: Early add of forced update requested cache variable
.. this is added in order to force the `forced_update_requested`
variable into the evaluation context variable cache. This is because in
the cases of early policy terminations, it's possible for the
`forced_update_requested` variable to never be cached, which means that
forced updates/installations will actually never trigger a re-eval of
the policies within the current evaluation context.

In the case of legacy DLC installations, it's critical to never depend
on the scheduled policy that is usually scheduled for a much later
timeout to kick off re-evaluation.

BUG=b:287912246
TEST=deploy update_engine + trigger DLC installation blocked on policy
ealier than forced update requested checks.

Change-Id: I72ec77f7a2395e092d4a823998a5f6f6f10c0601
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/4630611
Commit-Queue: Yuanpeng Ni‎ <yuanpengni@chromium.org>
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Yuanpeng Ni‎ <yuanpengni@chromium.org>
Auto-Submit: Jae Hoon Kim <kimjae@chromium.org>