update_engine: Invalidate updates if policy disabled
Currently if updates are disabled by the enterprise policy, updates
still can bypass the policy and get installed, if updates are
in-progress or already pending.

Extend the DeviceAutoUpdateDisabled policy and invalidate in-progress
or pending updates, if the enterprise policy disabled updates.
Also cancel the download action if the policy disables the updates
during update downloads.

BUG=b:275530794
TEST=unit tests
TEST=manual tests, WIP tast tests

Change-Id: I200e61b3be6ea5b43b6036a9e85f81ad7347db57
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/4730171
Commit-Queue: Artyom Chen <artyomchen@google.com>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Tested-by: Artyom Chen <artyomchen@google.com>