update_engine: FSI data on Rollback
For enterprise rollback, Goldeneye needs to make a decision whether to
send a rollback image to a device or not. This is based on FSI versions,
to ensure we do not send non-working versions to a device.

To help with that decision, give Goldeneye an idea of the FSI version at
production time of a device.

In the future, we will write the fsi_version as vpd entry after
production. If that is available, send that on a rollback request.

Until then, we can estimate the fsi version at production time using
activate date of the device. If fsi_version is not available send
activate_date instead.

BUG=b:301924476
TEST=FEATURE=test emerge-amd64-generic update_engine
TEST=Deploy on a device, configure vpd values and verify they are sent
  in update requests

Cq-Depend: 4902002
Change-Id: I4a82efaf30a0e7e40b7f6e9f01c06cee084fa7e6
Reviewed-on: https://chromium-review.googlesource.com/c/aosp/platform/system/update_engine/+/4893929
Tested-by: Miriam Polzer <mpolzer@google.com>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Miriam Polzer <mpolzer@google.com>